Refactor ChatInputButton components and improve UI consistency
Updated button paddings, sizes, titles, and hover states for better UX across `ToolsToggle`, `StreamingToggle`, and `LLMProviderToggle`. Removed redundant test cases and streamlined imports by relocating `LLMProviderToggle`. These changes enhance maintainability and visual harmony.
