+++ src/components/ChatInputButton/StreamingToggle.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
-import { CiStreamOn, CiStreamOff } from "react-icons/ci";
-
-// or alternatively:
-// import { TbPlugConnected, TbPlugConnectedX } from 'react-icons/tb';
+import { CiStreamOn, CiStreamOff } from 'react-icons/ci';
 
 interface StreamingToggleProps {
     isStreaming: boolean;
@@ -13,15 +10,15 @@ export const StreamingToggle: React.FC<StreamingToggleProps> = ({ isStreaming, o
     return (
         <button
             onClick={() => onToggle(!isStreaming)}
-            className={`p-2 rounded-lg transition-colors duration-200 ${
+            className={`p-1 rounded transition-colors duration-200 ${
                 isStreaming ? 'bg-blue-100 text-blue-600' : 'bg-gray-100 text-gray-600'
             } hover:bg-blue-50`}
             title={isStreaming ? "Streaming responses enabled" : "Streaming responses disabled"}
         >
             {isStreaming ? (
-                <CiStreamOn size={24} />
+                <CiStreamOn size={16} />
             ) : (
-                <CiStreamOff size={24} />
+                <CiStreamOff size={16} />
             )}
         </button>
     );
